Clean up calls to TableRow.set() in BigQueryTornados
No need to unbox the Integer/Long and rebox them. The
input PCollections are already typed correctly.

----Release Notes----
[]
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=102718218

+++ examples/src/main/java/com/google/cloud/dataflow/examples/cookbook/BigQueryTornadoes.java
@@ -101,8 +101,8 @@ public class BigQueryTornadoes {
     @Override
     public void processElement(ProcessContext c) {
       TableRow row = new TableRow()
-          .set("month", c.element().getKey().intValue())
-          .set("tornado_count", c.element().getValue().longValue());
+          .set("month", c.element().getKey())
+          .set("tornado_count", c.element().getValue());
       c.output(row);
     }
   }
